pyproject.toml
[tool.poetry]
name = "Waste-Bot"
version = "0.1.0"
description = ""
authors = ["hkn"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.5"
albumentations = "^0.5.2"
efficientnet-pytorch = "^0.7.1"
pandas = "^1.2.4"
Pillow = "^8.2.0"
torchvision = "^0.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
